requirements.txt
flask
alembic
sqlalchemy
python-dotenv
requests